requirements_brain.txt
streamlit==1.23.1
Pillow==9.5.0
numpy==1.24.3
torch==2.0.0
torchvision==0.15.1
opencv-python==4.6.0.66
matplotlib==3.5.3
plotly==5.15.0
#git+https://github.com/facebookresearch/detectron2.git
